Incorporate Xbox splash into autoexec
