+++ app-autoexec/main.c
@@ -2,6 +2,7 @@
 #include <stdlib.h>
 #include <stdio.h>
 #include <string.h>
+#include <math.h>
 
 #include "mach_defines.h"
 #include "sdk.h"
@@ -16,6 +17,12 @@ extern char _binary_switch_bg_png_end;
 // Followed by the tileset
 extern char _binary_switch_tileset_png_start;
 extern char _binary_switch_tileset_png_end;
+
+extern char _binary_xbox_bg_png_start;
+extern char _binary_xbox_bg_png_end;
+
+extern char _binary_xbox_tiles_png_start;
+extern char _binary_xbox_tiles_png_end;
 
 //Pointer to the framebuffer memory.
 uint8_t *fbmem;
@@ -55,6 +62,19 @@ uint32_t *GFXSPRITES = (uint32_t *)0x5000C000;
 #define TILE_SWITCH_RIGHT   0x24
 #define TILE_SWITCH_SUPERCON 0X01
 #define TILE_SWITCH_RED     0x10
+
+/////////////////////////////////////////////////////////////////////////////
+//
+//  Constants for splash screen inspired by Xbox One
+
+#define XBOX_X 7
+#define XBOX_Y 0
+
+#define TILE_XBOX    0x0
+
+//  End splash screen constants
+//
+/////////////////////////////////////////////////////////////////////////////
 
 //Borrowed this from lcd.c until a better solution comes along :/
 static void __INEFFICIENT_delay(int n) {
@@ -102,6 +122,27 @@ static inline void __tile_b_translate(int dx, int dy) {
 
 uint32_t counter60hz(void) {
 	return GFX_REG(GFX_VBLCTR_REG);
+}
+
+void gfx_set_xlate_val(int layer, int xcenter, int ycenter, float scale, float rot) {
+	float scale_inv=(1.0/scale);
+	float dx_x=cos(rot)*scale_inv;
+	float dx_y=-sin(rot)*scale_inv;
+	float dy_x=sin(rot)*scale_inv;
+	float dy_y=cos(rot)*scale_inv;
+	float start_x=-xcenter;
+	float start_y=-ycenter;
+
+	int i_dx_x=64.0*dx_x;
+	int i_dx_y=64.0*dx_y;
+	int i_dy_x=64.0*dy_x;
+	int i_dy_y=64.0*dy_y;
+	int i_start_x=(-start_x+start_x*dx_x-start_y*dx_y)*64.0;
+	int i_start_y=(-start_y+start_y*dy_y-start_x*dy_x)*64.0;
+
+	GFX_REG(GFX_TILEA_OFF)=(i_start_y<<16)+(i_start_x&0xffff);
+	GFX_REG(GFX_TILEA_INC_COL)=(i_dx_y<<16)+(i_dx_x&0xffff);
+	GFX_REG(GFX_TILEA_INC_ROW)=(i_dy_y<<16)+(i_dy_x&0xffff);
 }
 
 /////////////////////////////////////////////////////////////////////////////
@@ -255,6 +296,107 @@ void switch_splash() {
 
 /////////////////////////////////////////////////////////////////////////////
 //
+//  Called by main() to run splash screen inspired by Xbox One. This
+//  function is fully self-contained and can be main() for a standaline app.
+//  Be sure to copy the relevant constant #define above if doing so.
+
+void xbox_splash() {
+	//Allocate framebuffer memory
+	fbmem=malloc(320*512/2);
+
+	// Turn off top row LEDs
+	MISC_REG(MISC_LED_REG)=0;
+
+	//Set up the framebuffer address.
+	GFX_REG(GFX_FBADDR_REG)=((uint32_t)fbmem)&0xFFFFFF;
+	//We're going to use a pitch of 512 pixels, and the fb palette will start at 256.
+	GFX_REG(GFX_FBPITCH_REG)=(FB_PAL_OFFSET<<GFX_FBPITCH_PAL_OFF)|(512<<GFX_FBPITCH_PITCH_OFF);
+	//Blank out fb while we're loading stuff.
+	GFX_REG(GFX_LAYEREN_REG)=0;
+
+	//Load up the default tileset and font.
+	//ToDo: loading pngs takes a long time... move over to pcx instead.
+	printf("Loading tiles...\n");
+	int gfx_tiles_err = gfx_load_tiles_mem(GFXTILES, &GFXPAL[0], &_binary_xbox_tiles_png_start, (&_binary_xbox_tiles_png_end-&_binary_xbox_tiles_png_start));
+	printf("Tiles initialized err=%d\n", gfx_tiles_err);
+
+
+	//The IPL leaves us with a tileset that has tile 0 to 127 map to ASCII characters, so we do not need to
+	//load anything specific for this. In order to get some text out, we can use the /dev/console device
+	//that will use these tiles to put text in a tilemap. It uses escape codes to do so, see
+	//ipl/gloss/console_out.c for more info.
+	//Note that without the setvbuf command, no characters would be printed until 1024 characters are
+	//buffered.
+	console=fopen("/dev/console", "w");
+	setvbuf(console, NULL, _IOLBF, 1024); //make console line buffered
+	if (console==NULL) {
+		printf("Error opening console!\n");
+	}
+
+	//we tell it to start writing from entry 0.
+	//Now, use a library function to load the image into the framebuffer memory. This function will also set up the palette entries,
+	//PAL offset changes the colors that the 16-bit png maps to?
+	gfx_load_fb_mem(fbmem, &GFXPAL[FB_PAL_OFFSET], 4, 512, &_binary_xbox_bg_png_start, (&_binary_xbox_bg_png_end-&_binary_xbox_bg_png_start));
+
+	//Flush the memory region to psram so the GFX hw can stream it from there.
+	cache_flush(fbmem, fbmem+FB_WIDTH*FB_HEIGHT);
+
+	//Copied from IPL not sure why yet
+	GFXPAL[FB_PAL_OFFSET+0x100]=0x00ff00ff; //Note: For some reason, the sprites use this as default bgnd. ToDo: fix this...
+	GFXPAL[FB_PAL_OFFSET+0x1ff]=0x40ff00ff; //so it becomes this instead.
+
+	//This makes sure not to read button still pressed from badge menu selection
+	__button_wait_for_release();
+
+	//Set map to tilemap B, clear tilemap, set attr to 0
+	//Not sure yet what attr does, but tilemap be is important as it will have the effect of layering
+	//on top of our scrolling game
+	fprintf(console, "\0331M\033C\0330A\n");
+	//Note that without the newline at the end, all printf's would stay in the buffer.
+
+
+	//Clear both tilemaps
+	memset(GFXTILEMAPA,0,0x4000);
+	memset(GFXTILEMAPB,0,0x4000);
+	//Clear sprites that IPL may have loaded
+	memset(GFXSPRITES,0,0x4000);
+
+	/********************************************************************************
+	 * Put your user code in there, return when it's time to exit back to bage menu *
+	 * *****************************************************************************/
+	uint8_t x_offset = 0;
+	uint8_t y_offset = 0;
+	uint8_t tile_index = 0;
+
+	// Place tiles for Xbox sphere (actually circle since we're 2D)
+	for (uint8_t x = 0; x < 16; x++) {
+		x_offset = XBOX_X + x;
+		for (uint8_t y = 0; y < 16; y++) {
+			y_offset = XBOX_Y + y;
+			tile_index = TILE_XBOX + x + 0x10*y;
+			__tile_a_set(x_offset, y_offset, tile_index);
+		}
+	}
+
+	double scale = 6.3;
+
+	gfx_set_xlate_val(0, 240, 128, scale, 0);
+
+	// Tiles are set up, we can now enable layers
+	 GFX_REG(GFX_LAYEREN_REG)=GFX_LAYEREN_FB|GFX_LAYEREN_TILEA;
+
+	for (uint8_t i = 0; i < 60; i++) {
+		scale -= 0.1;
+		gfx_set_xlate_val(0, 240, 128, scale, 0);
+		__INEFFICIENT_delay(1);
+	}
+
+	// Logo complete, allow admiration for a short time before exiting.
+	__INEFFICIENT_delay(750);
+}
+
+/////////////////////////////////////////////////////////////////////////////
+//
 //  Randomly choose one of available splash screens to display
 
 void main(int argc, char **argv) {
@@ -280,6 +422,7 @@ void main(int argc, char **argv) {
 		case 3:
 		case 4:
 		case 5:
+			xbox_splash();
 			break;
 	}
 }
